Took out checkJwt and checkIfAdmin from all files and import them from middleware instead for reusability and cleaner code
