+++ backend-app/src/presentation-layer/admin-router-api.js
@@ -1,32 +1,9 @@
 var express = require('express')
-const jwtAuthz = require('express-jwt-authz');
-const jwt = require('express-jwt');
-const jwksRsa = require('jwks-rsa');
+const checkJwt = require('./middleware/jwt.middleware')
+const checkIfAdmin = require('./middleware/admin.middleware')
 
 const AdminCognitoService = require('./services/cognito.admin.config')
 
-// Put in .env variable?
-const checkIfAdmin = jwtAuthz(['Admins'], { customScopeKey: 'cognito:groups' });
-const region = 'eu-west-1';
-const adminUserPoolId = process.env.ADMIN_POOL;
-
-const checkJwt = jwt({
-    // Dynamically provide a signing key
-    // based on the kid in the header and 
-    // the signing keys provided by the JWKS endpoint.
-    secret: jwksRsa.expressJwtSecret({
-        cache: true,
-        rateLimit: true,
-        jwksRequestsPerMinute: 5,
-        jwksUri: `https://cognito-idp.${region}.amazonaws.com/${adminUserPoolId}/.well-known/jwks.json`,
-        // jwksUri: `https://dev-t3vri3ge.us.auth0.com/.well-known/jwks.json`
-    }),
-
-    // Validate the audience and the issuer.
-    // audience: 'flexicharge.app',
-    issuer: [`https://dev-t3vri3ge.us.auth0.com/`, `https://cognito-idp.eu-west-1.amazonaws.com/${adminUserPoolId}`],
-    algorithms: ['RS256']
-});
 
 module.exports = function () {
     const router = express.Router()
@@ -160,7 +137,7 @@ module.exports = function () {
     })
     router.post('/', checkJwt, checkIfAdmin, function (req, res) {
         const { username, password } = req.body;
-        
+
         cognito.createAdmin(username, password)
             .then(result => {
                 if (result.statusCode === 201) {
