+++ backend-app/src/presentation-layer/middleware/admin.middleware.js
@@ -1,0 +1,5 @@
+const jwtAuthz = require('express-jwt-authz');
+
+const checkIfAdmin = jwtAuthz(['Admins'], { customScopeKey: 'cognito:groups' });
+
+module.exports = checkIfAdmin
