+++ backend-app/src/presentation-layer/middleware/jwt.middleware.js
@@ -1,0 +1,23 @@
+const jwt = require('express-jwt');
+const jwksRsa = require('jwks-rsa');
+
+const checkJwt = jwt({
+    // Dynamically provide a signing key
+    // based on the kid in the header and 
+    // the signing keys provided by the JWKS endpoint.
+    secret: jwksRsa.expressJwtSecret({
+        cache: true,
+        rateLimit: true,
+        jwksRequestsPerMinute: 5,
+        jwksUri: `https://cognito-idp.${process.env.AWS_REGION}.amazonaws.com/${process.env.ADMIN_POOL}/.well-known/jwks.json`,
+        // jwksUri: `https://dev-t3vri3ge.us.auth0.com/.well-known/jwks.json`
+    }),
+
+    // Validate the audience and the issuer.
+    // audience: 'flexicharge.app',
+    issuer: [`https://dev-t3vri3ge.us.auth0.com/`, `https://cognito-idp.eu-west-1.amazonaws.com/${process.env.ADMIN_POOL}`],
+    algorithms: ['RS256']
+});
+
+
+module.exports = checkJwt
